Reading nelgt and nelgv from nelgv from re2
When `nelgt` and `nelgv` are parsed from the re2 file, there is a type mismatch.  They are declared as `long long int`
```
//host index data type
#if 0
#define hlong int
#define MPI_HLONG MPI_INT
#define hlongFormat "%d"
#define hlongString "int"
#else
#define hlong long long int
#define MPI_HLONG MPI_LONG_LONG_INT
#define hlongFormat "%lld"
#define hlongString "long long int"
#endif
```
However, the `sscanf` function read them as `int`.  This led to some data corruption, which was apparent in the generated SIZE file.  

This PR fixes the issue by reading them as `long long int`