Bugfix: Fix restarted GMRES code when using solution projection.
Previously, the rhs used to recompute the residual was incorrect
whenever solution projection was being used.
This caused cases to converge to the wrong pressure value
whenever GMRES had to restart.

This bug did not affect runs using PCG or where solution projection was not used.
The CI tests did not catch this, as all of the pressure solves converged
prior to needing to be restarted.