Enable new tavg plugin for CHT use
Since v23, the tavg module allows to specify the fields (and pointwise
products) to be averaged. However the kernel always operated on
the size of the vmesh only, i.e. when passing the temperature (or
any other field defined on the tmesh), the values in the solid were
ignored.

With this change, one may specify which mesh the kernel operates
on alongside each vector of fields.